Re-implement 'Modify storage' target to open cockpit storage in-page
This commit introduces a new user workflow for storage configuration in the installer.
The Cockpit storage editor, is replacing the blivet-gui and is now presented inside an iframe.
The communication between Anaconda and Cockpit is doen via localStorage.
The API which is still developed is defined in
[Cockpit Anaconda Documentation](https://github.com/cockpit-project/cockpit/blob/main/doc/anaconda.md).

The updated user workflow is as follows:
- Access the storage editor via the 'Modify storage' button.
- Requirements and recommendations for manual partitioning are detailed in the right sidebar.
- Upon completing storage manipulations, users can return to the installer wizard using the 'Return to installation' button.
- Upon clicking the button, the system checks for the possibility of two scenarios:
  1) Utilize manually configured partitioning.
  2) Utilize free space for automatic partitioning.
- The dialog prioritizes the previous order, creating the manual partitioning object for scenario 1,
and returning to the main screen. The selected scenario aligns with the user's presumed intention based on their storage edits.