tests: stabilization: wait for the disk to get detected after re-scan
I noticed a few tests failing on the disk selection screen. My
assumption is that re-scanning takes long and we do not wait for the
task to finish before proceeding.
I have attempted to wait for the re-scan task to finish after calling
`dbus_call_scan_devices` but there seems to be some bug with the thread
numbering which we need to debug.